pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.1.0a5", "cython", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "SCORE"
version = "0.1.0"
description = "Python package combining methods of embedding and clustering single-cell Hi-C data"
readme = "README.md"
authors = ["Dylan Plummer <dmp131@case.edu>"]
license = "MIT"
repository = "https://github.com/dylan-plummer/SCORE"
homepage = "https://github.com/dylan-plummer/SCORE"

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  #! Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  #! Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
]


[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
score = "score.__main__:app"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
importlib_metadata = {version = ">=1.0,<2.0", python = "<3.8"}
rich = "^13.0.1"
scHiCTools = { git = "https://github.com/dylan-plummer/scHiCTools.git", rev = "2501c4ea90b30759c0be26c17d4e862ebe72a77e" }
higashi = { git = "https://github.com/dylan-plummer/Higashi.git", branch = "main" }
fast-higashi = { git = "https://github.com/dylan-plummer/Fast-Higashi.git", branch = "main" }
joblib = "^1.1.0"
matplotlib = "^3.5.1"
numpy = { version = ">=1.22.2", python = ">=3.8, <3.10" }
pandas = ">=1.1.5,<2.0.0"
tqdm = "^4.62.3"
tensorflow = ">=2.14.0"
jax = ">=0.4.10"
scipy = "^1.7.3"
scikit-learn = "^1.0.2"
seaborn = "^0.12.2"
umap-learn = "^0.5.2"
munkres = "^1.1.4"
networkx = "^2.6.3"
cooler = "^0.9.1"
anndata = "0.8.0"
scanpy = "^1.8.2"
scvi-tools = ">=1.0.3"
wandb = ">=0.15.12"
torch = "<2.0.0"
tensorflow-probability = ">=0.22.0"
pingouin = "^0.5.1"
swifter = "^1.1.3"
plotly = "^5.11.0"
cython = "^0.29.33"
igraph = "^0.10.3"
louvain = "^0.8.0"
leidenalg = "^0.9.1"

[tool.poetry.dev-dependencies]
bandit = "^1.7.1"
black = {version = "^23.1a1", allow-prereleases = true}
darglint = "^1.8.1"
isort = {extras = ["colors"], version = "^5.10.1"}
mypy = "^0.910"
mypy-extensions = "^0.4.3"
pre-commit = "^2.15.0"
pydocstyle = "^6.1.1"
pylint = "^2.11.1"
pytest = "^7.2.1"
pyupgrade = "^2.29.1"
safety = "^2.2.0"
coverage = "^7.1.0"
coverage-badge = "^1.1.0"
pytest-html = "^3.1.1"
pytest-cov = "^4.0.0"
cython = "^0.29.33"

[tool.black]
# https://github.com/psf/black
target-version = ["py39"]
line-length = 88
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
    | examples
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 39
line_length = 88

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = 3.7
exclude = 'score/methods/scHiCTools/build/'
pretty = true
show_traceback = true
color_output = true

allow_untyped_globals = true
allow_redefinition = true
local_partial_types = true
check_untyped_defs = false
disallow_any_generics = false
disallow_incomplete_defs = false
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true



[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--dataset_config=flyamer.json",
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]

[tool.coverage.run]
source = ["tests"]

[coverage.paths]
source = "score"

[coverage.run]
branch = true

[coverage.report]
fail_under = 50
show_missing = true